random humanoid prototype into shared
